Fixed a typo on CsvDataManager that caused the calc to increase exponentially.
Added the class Parameters.cs for storing user entered inputs on Program.cs. That class will be serialized to a file and loaded on next startup instead of asking all the inputs again. I plan on using this class everywhere to not depend on so many class variables.

The ParametersFilePath will be entered by the user and stored into app.settings and loaded on next execution automatically.